修改

+++ TiHMS/CommonModule/src/main/java/tiquartet/CommonModule/vo/HotelVO.java
@@ -4,7 +4,7 @@ public class HotelVO {
     public int hotelID;
     public String name;
     public int level;
-    public int averagegrade;
+    public double averagegrade;
     public String  cityName;
     public String  businessName;
     public int lowestPrice;
